Tested save var and get var successful.
Key things to remember when calling the sys function, when calling get var, in user program, the memory must be allocated, it is preferable to use things like 'char varname[1000];' to initialize variable, so throughout the code it is consistent.  This would guarantee the memory is allocated and writable to memory space.

Feature added this commit:
*In save var , check if var name already in the list, replace vardef for the varname already exists instead of adding new one.
*Added P3_DEBUG definition so we can quickly change build for release build.

Minor Changes:
Modified length calculation in NextVariable function.